+++ include/QF/sys.h
@@ -97,7 +97,7 @@ void Sys_Init_Cvars (void);
 //
 // memory protection
 //
-void Sys_MakeCodeWriteable (unsigned long startaddr, unsigned long length);
+void Sys_MakeCodeWriteable (uintptr_t startaddr, size_t length);
 void Sys_PageIn (void *ptr, int size);
 
 //
